csgcca: remove -fanalyzer from compiler invocation
When we run `gcc -fanalyzer` in a separate process, there is no need to
pass `-fanalyzer` to the real compiler that produces binaries despite
the build system of an RPM package asks for it.

Besides it wastes resources, it could also break the build in case the
cswrap timeout is triggered for the real compiler (because cswrap treats
it as analyzer when it sees the `-fanalyzer` flag).

This problem was encountered while trying to scan gnutls-3.7.0-2.el9.